+++ src/backend/dnnl/common.cpp
@@ -303,7 +303,7 @@ void fill_layout_info(impl::logical_tensor_t *lt, const memory::desc &td) {
 
         impl::utils::optional<size_t> layout_id
                 = dnnl_backend::get_singleton().set_mem_desc(td);
-        lt->layout.layout_id = static_cast<int64_t>(layout_id.value());
+        lt->layout.layout_id = layout_id.value();
         lt->layout_type = impl::layout_type::opaque;
     }
 }
@@ -321,9 +321,8 @@ void fill_layout_info(
                     static_cast<impl::data_type_t>(td.data.data_type));
         }
 #endif // DNNL_GRAPH_LAYOUT_DEBUG
-
-        val->set_layout_id(static_cast<int64_t>(
-                dnnl_backend::get_singleton().set_mem_desc(td).value()));
+        val->set_layout_id(
+                dnnl_backend::get_singleton().set_mem_desc(td).value());
     }
 }
