+++ src/interface/logical_tensor.cpp
@@ -34,7 +34,7 @@ size_t logical_tensor_wrapper::size() const {
 
         return max_size * data_type_size();
     } else if (is_opaque()) {
-        size_t layout_id = static_cast<size_t>(lt->layout.layout_id);
+        size_t layout_id = lt->layout.layout_id;
         auto backend = backend_registry::get_singleton().get_registered_backend(
                 layout_id);
 
@@ -42,8 +42,8 @@ size_t logical_tensor_wrapper::size() const {
         // the encoded backend id from the layout id. Because each backend is
         // invisble about backend id for simplifing the backend integration
         logical_tensor_t new_lt = *lt;
-        new_lt.layout.layout_id = static_cast<int64_t>(
-                backend_registry::extract_layout_id(layout_id));
+        new_lt.layout.layout_id
+                = backend_registry::extract_layout_id(layout_id);
         return backend->get_mem_size(new_lt);
     } else {
         return (size_t)-1;
@@ -121,8 +121,8 @@ bool logical_tensor_wrapper::is_similar(const logical_tensor_t &lhs,
         // call backend to check whether a opaque layout is actually
         // equal to a strided layout in backend's perspective
         size_t layout_id = lhs.layout_type == layout_type::opaque
-                ? static_cast<size_t>(lhs.layout.layout_id)
-                : static_cast<size_t>(rhs.layout.layout_id);
+                ? lhs.layout.layout_id
+                : rhs.layout.layout_id;
         auto backend = backend_registry::get_singleton().get_registered_backend(
                 layout_id);
 
@@ -131,8 +131,8 @@ bool logical_tensor_wrapper::is_similar(const logical_tensor_t &lhs,
         // invisble about backend id for simplifying the backend integration
         logical_tensor_t new_lt
                 = lhs.layout_type == layout_type::opaque ? lhs : rhs;
-        new_lt.layout.layout_id = static_cast<int64_t>(
-                backend_registry::extract_layout_id(layout_id));
+        new_lt.layout.layout_id
+                = backend_registry::extract_layout_id(layout_id);
 
         return lhs.layout_type == layout_type::opaque
                 ? backend->compare_logical_tensor(new_lt, rhs)
@@ -159,8 +159,7 @@ size_t logical_tensor_wrapper::hash() const noexcept {
                         seed, this->strides(), nd);
             break;
         case layout_type::opaque:
-            seed = utils::hash_combine(
-                    seed, static_cast<size_t>(this->layout_id()));
+            seed = utils::hash_combine(seed, this->layout_id());
             break;
         default: assertm(false, "unknown layout_type");
     }
